refactor: 💡 Fix problems found by PR #16, part 1
Changes not include the rc data build process.
